fix(dashboards): dashboard page 500'd in prod (DuckDB in SSR); split lib + guard
PR #103 turned the dashboard page into a SERVER component that imported
`dashboardViewData` from `@/lib/dashboards`, which transitively pulls
`@/lib/malloy` → @duckdb/node-api. A Next PAGE render function can't load
libduckdb.so (it isn't traceable into page bundles, and outputFileTracingIncludes
can't target a page — verified on Next 16), so every ecommerce dashboard 500'd in
prod with "libduckdb.so: cannot open shared object file". This is the same trap
as the ltool page (reference: PR #79/#80).

Structural fix (no fragile lazy-import invariant):

- Split `@/lib/dashboards` into:
    • meta.ts   — DB-only (getDashboard, isCustomDashboard, list*). The ONLY
                  dashboard module a page may import. No path to Malloy/DuckDB.
    • engine.ts — the Malloy query/introspection work, plain STATIC imports of
                  @/lib/malloy. Imported ONLY by API routes (run/frame/view).
  The `@/lib/dashboards` barrel re-exports meta only, so importing it from a page
  is always safe; needing the engine is an explicit `/engine` import.

- The page now does a DB-only read to branch custom (sandboxed iframe) vs
  tag-only (in-page). Tag-only info + given specs (which need Malloy) are served
  by a new /api/dashboards/[datasetId]/[name]/view route and fetched by
  TagOnlyDashboard client-side — Malloy runs in the API route, where DuckDB is
  bundled.

- scripts/check-page-no-duckdb.mjs (wired into preflight): walks the STATIC
  import graph from every app/**/page.tsx and fails the build if it reaches a
  DuckDB package. Ignores dynamic import() and `import type` (neither 500s). Turns
  this whole class of bug into a red preflight check instead of a silent prod 500.

Verified: guard passes clean (11 pages) and catches a planted regression with the
exact path; typecheck + lint clean; prod-faithful standalone build with
libduckdb.so renamed → dashboard pages 200 (were 500); /view returns data; full
in-page render confirmed on localhost:3000 against the prod-fork DB.
